docs: enable theme-provider
